Refactored and cleaned up the assets folder
