+++ src/assets/styles/backgroundPallete.ts
