--- src/assets/styles/styles.d.ts
@@ -1,4 +1,0 @@
-declare module '*.css' {
-  const content: { [className: string]: string };
-  export default content;
-}
